+++ R/CoPro-package.R
@@ -1,0 +1,6 @@
+#' CoPro package native routines
+#'
+#' @useDynLib CoPro, .registration = TRUE
+#' @importFrom Rcpp evalCpp
+#' @keywords internal
+"_PACKAGE"